# docs/contracts/openapi.phase1.yaml
openapi: 3.1.0
info:
  title: CraftFlow Studio Phase 1 API
  version: 1.2.0-phase1ab
  description: Prompt Maker and Website Maker contracts with quality/regeneration/history endpoints.
servers:
  - url: https://api.craftflow.studio
security:
  - bearerAuth: []
paths:
  /v1/health:
    get:
      summary: Health check
      security: []
      responses:
        '200': { description: Service healthy }

  /v1/auth/session:
    post:
      summary: Create session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
    delete:
      summary: End session
      responses:
        '204': { description: Session ended }

  /v1/projects:
    get:
      summary: List projects
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }

  /v1/projects/{projectId}:
    get:
      summary: Get project by id
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update project
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProjectRequest' }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    delete:
      summary: Archive project
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      responses:
        '204': { description: Project archived }

  /v1/projects/{projectId}/generations:
    get:
      summary: List project generation history
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      responses:
        '200':
          description: Generation history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Generation' }

  /v1/prompt-maker/blueprint/generations:
    post:
      summary: Submit Business Blueprint prompt generation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBlueprintPromptRequest' }
      responses:
        '202':
          description: Prompt generation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationAcceptedResponse' }

  /v1/prompt-maker/digital-product/generations:
    post:
      summary: Submit Digital Product prompt generation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDigitalProductPromptRequest' }
      responses:
        '202':
          description: Prompt generation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationAcceptedResponse' }

  /v1/prompt-maker/blueprint/quality:
    post:
      summary: Score blueprint prompt quality
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBlueprintPromptRequest' }
      responses:
        '200': { description: Quality report }

  /v1/prompt-maker/digital-product/quality:
    post:
      summary: Score digital product prompt quality
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDigitalProductPromptRequest' }
      responses:
        '200': { description: Quality report }

  /v1/prompt-maker/blueprint/regenerate-section:
    post:
      summary: Build blueprint section regeneration prompt
      requestBody:
        required: true
      responses:
        '200': { description: Regeneration prompt }

  /v1/prompt-maker/digital-product/regenerate-section:
    post:
      summary: Build digital product section regeneration prompt
      requestBody:
        required: true
      responses:
        '200': { description: Regeneration prompt }

  /v1/website-maker/generations:
    post:
      summary: Submit Website Maker generation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWebsiteGenerationRequest' }
      responses:
        '202':
          description: Website generation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationAcceptedResponse' }

  /v1/generations/{generationId}:
    get:
      summary: Get generation by id
      parameters: [{ $ref: '#/components/parameters/GenerationId' }]
      responses:
        '200':
          description: Generation record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Generation' }
        '404': { $ref: '#/components/responses/NotFound' }

  /v1/usage/summary:
    get:
      summary: Get token and cost summary
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageSummary' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    GenerationId:
      name: generationId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    CreateSessionRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    SessionResponse:
      type: object
      required: [token, expires_at, user_id]
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
        user_id: { type: string, format: uuid }
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, maxLength: 1000 }
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, maxLength: 1000 }
        status: { type: string, enum: [active, archived] }
    Project:
      type: object
      required: [id, user_id, name, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateBlueprintPromptRequest:
      type: object
      required: [project_id, tool_target, questionnaire]
      properties:
        project_id: { type: string, format: uuid }
        tool_target: { type: string, enum: [chatgpt, claude, gemini] }
        questionnaire: { type: object, additionalProperties: true }
        idempotency_key: { type: string, minLength: 8, maxLength: 128 }
    CreateDigitalProductPromptRequest:
      type: object
      required: [project_id, tool_target, questionnaire]
      properties:
        project_id: { type: string, format: uuid }
        tool_target: { type: string, enum: [chatgpt, claude, gemini] }
        questionnaire: { type: object, additionalProperties: true }
        idempotency_key: { type: string, minLength: 8, maxLength: 128 }
    CreateWebsiteGenerationRequest:
      type: object
      required: [project_id, product_type, audience, color_theme, store_name]
      properties:
        project_id: { type: string, format: uuid }
        product_type: { type: string, maxLength: 100 }
        audience: { type: string, maxLength: 200 }
        color_theme: { type: string, maxLength: 100 }
        store_name: { type: string, maxLength: 160 }
        has_shopify_account: { type: boolean }
        idempotency_key: { type: string, minLength: 8, maxLength: 128 }
    GenerationAcceptedResponse:
      type: object
      required: [generation_id, status]
      properties:
        generation_id: { type: string, format: uuid }
        status: { type: string, enum: [queued, processing] }
    Generation:
      type: object
      required: [id, request_type, project_id, request_id, status, provider_name, model_name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        request_type: { type: string, enum: [prompt_blueprint, prompt_product, website_maker] }
        project_id: { type: string, format: uuid }
        request_id: { type: string, format: uuid }
        status: { type: string, enum: [queued, processing, completed, failed, blocked] }
        provider_name: { type: string }
        model_name: { type: string }
        output_text: { type: string }
        error_code: { type: string }
        error_message: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UsageSummary:
      type: object
      required: [total_requests, total_tokens, total_cost_usd]
      properties:
        total_requests: { type: integer, minimum: 0 }
        total_tokens: { type: integer, minimum: 0 }
        input_tokens: { type: integer, minimum: 0 }
        output_tokens: { type: integer, minimum: 0 }
        total_cost_usd: { type: number, format: double, minimum: 0 }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }